MASSIVE refactor to the App.jsx
